+++ src/bfabric/experimental/app_interface/app_runner/_spec.py
@@ -12,3 +12,5 @@ class CommandsSpec(BaseModel):
 class AppSpec(BaseModel):
     version: list[str] = []
     commands: CommandsSpec
+    # Note: While we use the old submitter, this is still necessary
+    reuse_default_resource: bool = True