+++ docs/changelog.md
@@ -20,6 +20,7 @@ Versioning currently follows `X.Y.Z` where
 - (internal) migrate to src layout
 - (experimental) the former `process` step of the app runner has been split into a `process` and `collect` step where,
     the collect step is responsible for generating the `output.yml` file that will then be used to register the results.
+- (experimental) app runner apps by default reuse the default resource
 
 ## \[1.13.7\] - 2024-09-17